# .postman/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "description": "Returns a welcome message",
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Welcome message returned successfully"
          }
        },
        "summary": "Hello endpoint",
        "tags": [
          "App"
        ]
      }
    },
    "/health": {
      "get": {
        "description": "Check the health status of the application",
        "operationId": "AppController_healthCheck",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Application is healthy"
          }
        },
        "summary": "Health check",
        "tags": [
          "App"
        ]
      }
    },
    "/users": {
      "get": {
        "description": "Retrieve a list of all users",
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of users returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "description": "Create a new user",
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Create user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Retrieve a single user by their ID",
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found and returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "description": "Update an existing user by ID",
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Update user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "description": "Delete a user by ID",
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Delete user",
        "tags": [
          "Users"
        ]
      }
    },
    "/items": {
      "get": {
        "description": "Retrieve a list of all items",
        "operationId": "ItemsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of items returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get all items",
        "tags": [
          "Items"
        ]
      },
      "post": {
        "description": "Create a new item",
        "operationId": "ItemsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Create item",
        "tags": [
          "Items"
        ]
      }
    },
    "/items/search": {
      "get": {
        "description": "Search items by query string",
        "operationId": "ItemsController_search",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search query string",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Search items",
        "tags": [
          "Items"
        ]
      }
    },
    "/items/{id}": {
      "get": {
        "description": "Retrieve a single item by its ID",
        "operationId": "ItemsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Item ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item found and returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "404": {
            "description": "Item not found"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get item by ID",
        "tags": [
          "Items"
        ]
      },
      "patch": {
        "description": "Update an existing item by ID",
        "operationId": "ItemsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Item ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "404": {
            "description": "Item not found"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Update item",
        "tags": [
          "Items"
        ]
      },
      "delete": {
        "description": "Delete an item by ID",
        "operationId": "ItemsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Item ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "404": {
            "description": "Item not found"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Delete item",
        "tags": [
          "Items"
        ]
      }
    }
  },
  "info": {
    "title": "NestJS API",
    "description": "REST API documentation",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authentication"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john@example.com"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "createdAt"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "The email address of the user",
            "example": "john.doe@example.com"
          }
        },
        "required": [
          "name",
          "email"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "The email address of the user",
            "example": "john.doe@example.com"
          }
        }
      },
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Name of the item",
            "example": "Widget"
          },
          "description": {
            "type": "string",
            "description": "Description of the item",
            "example": "A useful widget"
          },
          "price": {
            "type": "number",
            "description": "Price in dollars",
            "example": 19.99
          },
          "quantity": {
            "type": "number",
            "description": "Quantity in stock",
            "example": 100
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "price",
          "quantity",
          "createdAt"
        ]
      },
      "CreateItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the item",
            "example": "Widget"
          },
          "description": {
            "type": "string",
            "description": "The description of the item",
            "example": "A useful widget for everyday tasks"
          },
          "price": {
            "type": "number",
            "description": "The price of the item in dollars",
            "example": 19.99
          },
          "quantity": {
            "type": "number",
            "description": "The quantity available in stock",
            "example": 100
          }
        },
        "required": [
          "name",
          "description",
          "price",
          "quantity"
        ]
      },
      "UpdateItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the item",
            "example": "Widget"
          },
          "description": {
            "type": "string",
            "description": "The description of the item",
            "example": "A useful widget for everyday tasks"
          },
          "price": {
            "type": "number",
            "description": "The price of the item in dollars",
            "example": 19.99
          },
          "quantity": {
            "type": "number",
            "description": "The quantity available in stock",
            "example": 100
          }
        }
      }
    }
  }
}